Edit feestatements amount

+++ database/migrations/2021_06_22_161412_create_feestatements-table.php
@@ -19,7 +19,7 @@ class CreateFeestatementsTable extends Migration
             $table->dateTime('date');
             $table->string('document_number')->nullable();
             $table->string('document_type')->nullable();
-            $table->integer('amount');
+            $table->double('amount');
             $table->string('type');
             $table->timestamps();
